Agentic_AI_System/backend-api/pyproject.toml
[project]
name = "backend-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Elric Ettmüller", email = "elric.ettmueller@hm.edu"},
    {name = "Azim Alizada", email = "azim.alizada@hm.edu"},
    {name = "Maximilian Weicht", email = "maximilian.weicht@hm.edu"},
    {name = "Alessandro Fisslinger", email = "alessandro.fisslinger@hm.edu"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "poetry-core (>=2.1.3,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "crewai (>=0.121.1,<0.122.0)",
    "langchain-core (>=0.3.60,<0.4.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "langchain-ollama (>=0.1.0,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "PyYAML (>=6.0,<7.0)",
    "litellm (>=1.15.0,<2.0.0)",
    "easyocr (>=1.6.1,<2.0.0)",
    "pdfminer.six (>=20221105)",
    "sentence-transformers (>=2.2.2,<3.0.0)",
    "keybert (>=0.5.0,<1.0.0)",
    "language-tool-python (>=1.0.0,<2.0.0)",
    "python-multipart (>=0.0.5,<0.1.0)",
    "PyMuPDF (==1.23.7)",
    "numpy (>=1.24.0,<2.0.0)",
    "pymongo (>=4.13.2,<5.0.0)",
    "pillow-heif (>=0.15.0,<0.16.0)",
    "Pillow (>=9.0.0,<10.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
